Dependency class moved from Core to DataEditor
