+++ DataEditor/DataEditor.vcxproj.filters
@@ -30,6 +30,9 @@
     <ClInclude Include="EditableField.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="Dependency.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="EditReader.cpp">
@@ -45,6 +48,9 @@
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="EditableField.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="Dependency.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
