+++ MergeReport/MergeReportBase.h
@@ -1,6 +1,7 @@
 #pragma once
 //#include "../Core/Compiler.h"
 #include "../Core/rdrun.h"
+#include "ConstListEl.h"
 
 class Compiler;
